Revert of Split content_shell dependencies for layout tests. (https://codereview.chromium.org/221993002/)
Reason for revert:
Speculative rollout due to Linux DEPS builder compilation failure: http://build.chromium.org/p/chromium.webkit/builders/WebKit%20Linux%20%28deps%29/builds/27968

Original issue's description:
> Split content_shell dependencies for layout tests.
>
> This CL is isolating the minimum dependencies required to make layout
> tests work in content_shell without requiring it to depend on the entire
> test_support_content_target.
> crrev.com/11275277 has introduced a dependency on test_support_content
> to cope with layout test. test_support_content, however, was intended
> for unit/browser-tests and adding its dependency to content_shell
> causes a lot of unnecessary sources to be linked in.
> The linker --gc-sections has hidden the problem so far, but that doesn't
> seem to be an option for arm64 (at least with the current toolchain).
>
>
> BUG=159847,354405
>
> Committed: https://src.chromium.org/viewvc/chrome?view=rev&revision=261741

TBR=jam@chromium.org,primiano@chromium.org
NOTREECHECKS=true
NOTRY=true
BUG=159847,354405

Review URL: https://codereview.chromium.org/226123003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@261759 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: d9d0a3f08ed4703b1f05bb9dc311670319adb4b9

+++ content/test/mock_webclipboard_impl.cc
@@ -102,7 +102,7 @@ blink::WebData MockWebClipboardImpl::readImage(
                         gfx::PNGCodec::FORMAT_BGRA,
 #endif
                         gfx::Size(bitmap.width(), bitmap.height()),
-                        static_cast<int>(bitmap.rowBytes()),
+                        bitmap.rowBytes(),
                         false /* discard_transparency */,
                         std::vector<gfx::PNGCodec::Comment>(),
                         &encoded_image);
